Rename namespaces from `LuhnDotNetTest.Extension` to `LuhnDotNetTest.Extensions` in test files
