+++ tests/Extensions/StringExtensionTest.cs
@@ -1,5 +1,4 @@
-#if NET8_0_OR_GREATER
-namespace LuhnDotNetTest.Extension;
+namespace LuhnDotNetTest.Extensions;
 
 using LuhnDotNet;
 using LuhnDotNet.Extensions;
@@ -7,7 +6,10 @@ using System;
 using System.Collections.Generic;
 using Xunit;
 
-public class ReadOnlySpanExtensionsTest
+/// <summary>
+/// Unit tests for the C# implementation of the Luhn algorithm.
+/// </summary>
+public class StringExtensionTest
 {
     /// <summary>
     /// Test data for AlphaNumericToNumeric method.
@@ -25,13 +27,13 @@ public class ReadOnlySpanExtensionsTest
     /// <summary>
     /// Tests the AlphaNumericToNumeric method.
     /// </summary>
-    /// <param name="input">Input ReadOnlySpan{char}</param>
+    /// <param name="input">Input string</param>
     /// <param name="expected">Expected output</param>
-    [Theory(DisplayName = "Converts an alphanumeric ReadOnlySpan<char> to a numeric ReadOnlySpan<char>")]
-    [MemberData(nameof(AlphaNumericToNumericData), MemberType = typeof(ReadOnlySpanExtensionsTest))]
+    [Theory(DisplayName = "Converts an alphanumeric string to a numeric string")]
+    [MemberData(nameof(AlphaNumericToNumericData), MemberType = typeof(StringExtensionTest))]
     public void AlphaNumericToNumeric_ValidInput_ShouldReturnExpectedResult(string input, string expected)
     {
-        Assert.Equal(expected.AsSpan(), input.AsSpan().AlphaNumericToNumeric());
+        Assert.Equal(expected, input.AlphaNumericToNumeric());
     }
 
     /// <summary>
@@ -42,27 +44,28 @@ public class ReadOnlySpanExtensionsTest
     /// invalid input string that contains non-alphanumeric characters. The test uses the Assert. Throws method from
     /// xUnit to check if the expected exception is thrown.
     /// </remarks>
-    [Fact(DisplayName = "Converts an invalid alphanumeric ReadOnlySpan<char> to a numeric ReadOnlySpan<char> to throw an exception")]
+    [Fact(DisplayName = "Converts an invalid alphanumeric string to a numeric string to throw an exception")]
     public void AlphaNumericToNumeric_InvalidInput_ThrowsInvalidCharacterException()
     {
-        Assert.Throws<InvalidCharacterException>(()=> "!@#$%^&*()".AsSpan().AlphaNumericToNumeric());
+        Assert.Throws<InvalidCharacterException>(()=> "!@#$%^&*()".AlphaNumericToNumeric());
     }
 
     /// <summary>
-    /// Tests that the AlphaNumericToNumeric method throws an ArgumentNullException when the input is empty or null.
+    /// Validates that the AlphaNumericToNumeric method throws an ArgumentNullException
+    /// when passed an empty or null string input.
     /// </summary>
-    [Fact(DisplayName = "Converts an empty alphanumeric ReadOnlySpan<char> to a numeric ReadOnlySpan<char> to throw an exception")]
+    [Fact(DisplayName = "Converts an empty alphanumeric string to a numeric string to throw an exception")]
     public void AlphaNumericToNumeric_EmptyInput_ThrowsArgumentNullException()
     {
-        Assert.Throws<ArgumentNullException>(() => "".AsSpan().AlphaNumericToNumeric());
-        Assert.Throws<ArgumentNullException>(() => ((string)null)!.AsSpan().AlphaNumericToNumeric());
+        Assert.Throws<ArgumentNullException>(() => "".AlphaNumericToNumeric());
+        Assert.Throws<ArgumentNullException>(() => ((string)null)!.AlphaNumericToNumeric());
     }
 
     /// <summary>
     /// Tests that the RemoveSeparators method processes input correctly by removing separators.
     /// </summary>
-    /// <param name="input">The input ReadOnlySpan{char} containing separators.</param>
-    /// <param name="expected">The expected ReadOnlySpan{char} after separators are removed.</param>
+    /// <param name="input">The input string containing separators.</param>
+    /// <param name="expected">The expected string after separators are removed.</param>
     [Theory]
     [InlineData("4444 5555 6666 8888", "4444555566668888")]
     [InlineData("6666/5555/7777/9123", "6666555577779123")]
@@ -71,17 +74,17 @@ public class ReadOnlySpanExtensionsTest
     [InlineData("0987654321", "0987654321")]
     public void RemoveSeparators_ValidInput_ReturnsExpectedResult(string input, string expected)
     {
-        Assert.Equal(expected.AsSpan(), input.AsSpan().RemoveSeparators());
+        Assert.Equal(expected, input.RemoveSeparators());
     }
 
     /// <summary>
-    /// Tests the RemoveSeparators method to ensure it throws an ArgumentNullException when provided with an empty or null input.
+    /// Tests the RemoveSeparators method to ensure it throws an ArgumentNullException
+    /// when the input string is null or empty.
     /// </summary>
-    [Fact(DisplayName = "Remove separators from an empty ReadOnlySpan<char> to throw an exception")]
+    [Fact(DisplayName = "Remove separators from an invalid string to throw an exception")]
     public void RemoveSeparators_EmptyInput_ThrowsArgumentNullException()
     {
-        Assert.Throws<ArgumentNullException>(() => "".AsSpan().RemoveSeparators());
-        Assert.Throws<ArgumentNullException>(() => ((string)null)!.AsSpan().RemoveSeparators());
+        Assert.Throws<ArgumentNullException>(() => "".RemoveSeparators());
+        Assert.Throws<ArgumentNullException>(() => ((string)null)!.RemoveSeparators());
     }
 }
-#endif
